RM Back-End/requirements.txt
aniso8601==9.0.1
attrs==21.4.0
beautifulsoup4==4.10.0
cachetools==5.0.0
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.12
click==8.0.4
cryptography==37.0.2
cv==1.0.0
cycler==0.11.0
dnspython==2.2.1
email-validator==1.1.3
Flask==2.0.3
Flask-Cors==3.0.10
Flask-JWT-Extended==4.4.0
Flask-Login==0.5.0
flask-mongoengine==1.0.0
flask-restx==0.5.1
Flask-WTF==1.0.0
fonttools==4.31.1
google==3.0.0
google-api-core==2.7.1
google-api-python-client==2.41.0
google-auth==2.6.0
google-auth-httplib2==0.1.0
google-auth-oauthlib==0.5.1
googleapis-common-protos==1.55.0
httplib2==0.20.4
idna==3.3
itsdangerous==2.1.1
Jinja2==3.0.3
joblib==1.1.0
jsonschema==4.4.0
kiwisolver==1.4.0
MarkupSafe==2.1.0
marshmallow==3.15.0
matplotlib==3.5.1
mongoengine==0.24.0
numpy==1.22.3
oauthlib==3.2.0
opencv-python==4.5.5.64
packaging==21.3
Pillow==9.0.1
protobuf==3.19.4
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pyheif==0.7.0
PyJWT==2.4.0
pymongo==4.0
pyparsing==3.0.7
pyrsistent==0.18.1
pytesseract==0.3.9
python-dateutil==2.8.2
python-dotenv==0.19.2
pytz==2021.3
requests==2.27.1
requests-oauthlib==1.3.1
rsa==4.8
schedule==1.1.0
scikit-learn==1.1.1
scipy==1.8.1
six==1.16.0
soupsieve==2.3.1
threadpoolctl==3.1.0
uritemplate==4.1.1
urllib3==1.26.8
Werkzeug==2.0.3
whatimage==0.0.3
WTForms==3.0.1
